Separate mock form nd back button into components
