+++ src/app/random-facts/page.js
@@ -1,10 +1,10 @@
 "use client";
 
-import { Title, Text, Button, Card, Stack } from "@mantine/core";
+import { Title, Text, Button, Card, Stack, Flex } from "@mantine/core";
+import BackButton from "@/components/BackButton";
 import { useRandomFactsQuery } from "@/hooks/useRandomFactsQuery";
 import { useCallback } from "react";
 import { useRouter } from "next/navigation";
-import { IconArrowLeft } from "@tabler/icons-react";
 
 export default function RandomFactsPage() {
   const router = useRouter();
@@ -17,19 +17,18 @@ export default function RandomFactsPage() {
   return (
     <>
       <Stack gap="md">
-        <Button
-          w="fit-content"
-          variant="outline"
-          leftSection={<IconArrowLeft size={16} />}
-          onClick={() => router.push("/")}
-        >
-          Back
-        </Button>
-
-        <Title order={1}>Random Facts API</Title>
+        <Flex justify="space-between" align="center" gap="md">
+          <BackButton />
+          <Title order={1}>Random Facts API</Title>
+        </Flex>
         <Text>This is a page that shows some random facts.</Text>
-        <Button color="blue" onClick={handleGetRandomFact} disabled={isLoading}>
-          {isLoading ? "Loading..." : error ? "Retry" : "Get Random Facts"}
+        <Button
+          color="blue"
+          onClick={handleGetRandomFact}
+          disabled={isLoading}
+          loading={isLoading}
+        >
+          {error ? "Retry" : "Get Random Facts"}
         </Button>
         <Card shadow="sm" padding="lg" radius="md" withBorder>
           <Text c={error ? "red" : "black"}>
